Fetch data from own API.
Use my own API instead of hitting the sites directly, let all the
processing be done on the backend.

Changed the data being diplayed since the API structure has changed.

Use watchers to determine when to fetch the data for a particular
component. Data should only be fetched when a particular component is
marked as active the very first time. This makes more sense than
fetching ALL the respective data when ALL the components have mounted.